Fix printf format specifiers for size_t arguments
Replace %d/%lu/%ld with (int) casts for size_t/size_type arguments
in fprintf/printf calls across mesh, CFD, FEA, and GUI files.
Also cast time_t to (long long) in ResultsMgr to match %lld.

Fixes -Wformat warnings on GCC/Clang (Linux, macOS) and
C4477 on MSVC (Windows).

+++ src/cfd_mesh/CfdMeshMgr.cpp
@@ -1789,7 +1789,7 @@ void CfdMeshMgrSingleton::WriteNASCART_Obj_Tri_Gmsh( const string &dat_fn, const
         {
             fprintf( fp, "# vspgeom v3\n" );
             fprintf( fp, "1\n" );  // Number of meshes.
-            fprintf( fp, "%d %d %d\n", allUsedPntVec.size(), nface, wakes.size() );
+            fprintf( fp, "%d %d %d\n", (int)allUsedPntVec.size(), nface, (int)wakes.size() );
 
             //==== Write Pnts ====//
             for ( int i = 0 ; i < ( int )allUsedPntVec.size() ; i++ )
